move danhMuc to data.js
